refactor: use supabaseClient everywhere
