+++ src/services/auth/authService.ts
@@ -1,4 +1,4 @@
-import { supabase } from "@/lib/supabase";
+import { supabase } from "@/lib/supabaseClient";
 import { logger } from "@/lib/logger";
 
 export interface User {
